move statetensor to tracing module
